requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.2
contourpy==1.1.1
cycler==0.12.1
DateTime==5.2
deepspeed==0.10.1
filelock==3.12.2
flops-profiler==0.1.2
fonttools==4.43.1
fvcore==0.1.5.post20221221
hjson==3.1.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
portalocker==2.7.0
psutil==5.9.5
py-cpuinfo==9.0.0
pydantic==1.10.12
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.13.0
six==1.16.0
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.8.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yacs==0.1.8
zipp==3.16.2
zope.interface==6.0
